requirements.txt
torch>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.13.4
pandas>=2.0.0
streamlit>=1.20.0
tqdm>=4.64.0
python-dotenv>=0.21.0
altair==4.2.2
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
entrypoints==0.4
executing==2.2.0
filelock==3.16.1
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.24.4
packaging==25.0
pandas==1.5.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.19.1
Pympler==1.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.20.1
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.22.0
tenacity==8.5.0
tokenizers==0.13.3
toml==0.10.2
toolz==1.0.0
torch==1.13.1+cpu
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.26.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.2
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.20.2
